Move SerializationFormattingPolicy to Implementation
