+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/SerializationFormattingPolicy.java
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.cosmos.models;
+package com.azure.cosmos.implementation;
 
 /**
  * The formatting policy associated with JSON serialization in the Azure Cosmos DB database service.
